+++ example/auth.py
@@ -1,12 +1,11 @@
 #!/usr/bin/python
-from __future__ import print_function
 from pyrad.client import Client
 from pyrad.dictionary import Dictionary
 import socket
 import sys
 import pyrad.packet
 
-srv = Client(server="localhost", secret=b"Kah3choteereethiejeimaeziecumi", dict=Dictionary("dictionary"))
+srv = Client(server="127.0.0.1", secret=b"Kah3choteereethiejeimaeziecumi", dict=Dictionary("dictionary"))
 
 req = srv.CreateAuthPacket(code=pyrad.packet.AccessRequest, User_Name="wichert")