+++ tests/testClient.py
@@ -1,14 +1,15 @@
+import select
 import socket
 import unittest
-import six
+from .mock import MockPacket
+from .mock import MockPoll
+from .mock import MockSocket
 from pyrad.client import Client
 from pyrad.client import Timeout
 from pyrad.packet import AuthPacket
 from pyrad.packet import AcctPacket
 from pyrad.packet import AccessRequest
 from pyrad.packet import AccountingRequest
-from pyrad.tests.mock import MockPacket
-from pyrad.tests.mock import MockSocket
 
 BIND_IP = "127.0.0.1"
 BIND_PORT = 53535
@@ -20,33 +21,33 @@ class ConstructionTests(unittest.TestCase):
 
     def testSimpleConstruction(self):
         client = Client(self.server)
-        self.failUnless(client.server is self.server)
+        self.assertTrue(client.server is self.server)
         self.assertEqual(client.authport, 1812)
         self.assertEqual(client.acctport, 1813)
-        self.assertEqual(client.secret, six.b(''))
+        self.assertEqual(client.secret, b'')
         self.assertEqual(client.retries, 3)
         self.assertEqual(client.timeout, 5)
-        self.failUnless(client.dict is None)
+        self.assertTrue(client.dict is None)
 
     def testParameterOrder(self):
         marker = object()
         client = Client(self.server, 123, 456, 789, "secret", marker)
-        self.failUnless(client.server is self.server)
+        self.assertTrue(client.server is self.server)
         self.assertEqual(client.authport, 123)
         self.assertEqual(client.acctport, 456)
         self.assertEqual(client.coaport, 789)
         self.assertEqual(client.secret, "secret")
-        self.failUnless(client.dict is marker)
+        self.assertTrue(client.dict is marker)
 
     def testNamedParameters(self):
         marker = object()
         client = Client(server=self.server, authport=123, acctport=456,
                       secret="secret", dict=marker)
-        self.failUnless(client.server is self.server)
+        self.assertTrue(client.server is self.server)
         self.assertEqual(client.authport, 123)
         self.assertEqual(client.acctport, 456)
         self.assertEqual(client.secret, "secret")
-        self.failUnless(client.dict is marker)
+        self.assertTrue(client.dict is marker)
 
 
 class SocketTests(unittest.TestCase):
@@ -56,6 +57,7 @@ class SocketTests(unittest.TestCase):
         self.orgsocket = socket.socket
         socket.socket = MockSocket
 
+
     def tearDown(self):
         socket.socket = self.orgsocket
 
@@ -63,7 +65,7 @@ class SocketTests(unittest.TestCase):
         self.client._SocketOpen()
         sock = self.client._socket
         self.client._SocketOpen()
-        self.failUnless(sock is self.client._socket)
+        self.assertTrue(sock is self.client._socket)
 
     def testBind(self):
         self.client.bind((BIND_IP, BIND_PORT))
@@ -74,6 +76,7 @@ class SocketTests(unittest.TestCase):
     def testBindClosesSocket(self):
         s = MockSocket(socket.AF_INET, socket.SOCK_DGRAM)
         self.client._socket = s
+        self.client._poll = MockPoll()
         self.client.bind((BIND_IP, BIND_PORT))
         self.assertEqual(s.closed, True)
 
@@ -119,7 +122,7 @@ class SocketTests(unittest.TestCase):
         self.client.timeout = 1
         packet = MockPacket(AccessRequest)
         self.assertRaises(Timeout, self.client._SendPacket, packet, 432)
-        self.failIf("Acct-Delay-Time" in packet)
+        self.assertFalse("Acct-Delay-Time" in packet)
 
     def testSingleAccountDelay(self):
         self.client.retries = 2
@@ -138,22 +141,25 @@ class SocketTests(unittest.TestCase):
     def testIgnorePacketError(self):
         self.client.retries = 1
         self.client.timeout = 1
-        self.client._socket = MockSocket(1, 2, six.b("valid reply"))
+        self.client._socket = MockSocket(1, 2, b'valid reply')
         packet = MockPacket(AccountingRequest, verify=True, error=True)
         self.assertRaises(Timeout, self.client._SendPacket, packet, 432)
 
     def testValidReply(self):
         self.client.retries = 1
         self.client.timeout = 1
-        self.client._socket = MockSocket(1, 2, six.b("valid reply"))
+        self.client._socket = MockSocket(1, 2, b'valid reply')
+        self.client._poll = MockPoll()
+        MockPoll.results = [(1, select.POLLIN)]
         packet = MockPacket(AccountingRequest, verify=True)
         reply = self.client._SendPacket(packet, 432)
-        self.failUnless(reply is packet.reply)
+        self.assertTrue(reply is packet.reply)
 
     def testInvalidReply(self):
         self.client.retries = 1
         self.client.timeout = 1
-        self.client._socket = MockSocket(1, 2, six.b("invalid reply"))
+        self.client._socket = MockSocket(1, 2, b'invalid reply')
+        MockPoll.results = [(1, select.POLLIN)]
         packet = MockPacket(AccountingRequest, verify=False)
         self.assertRaises(Timeout, self.client._SendPacket, packet, 432)
 
@@ -161,18 +167,18 @@ class SocketTests(unittest.TestCase):
 class OtherTests(unittest.TestCase):
     def setUp(self):
         self.server = object()
-        self.client = Client(self.server, secret=six.b('zeer geheim'))
+        self.client = Client(self.server, secret=b'zeer geheim')
 
     def testCreateAuthPacket(self):
         packet = self.client.CreateAuthPacket(id=15)
-        self.failUnless(isinstance(packet, AuthPacket))
-        self.failUnless(packet.dict is self.client.dict)
+        self.assertTrue(isinstance(packet, AuthPacket))
+        self.assertTrue(packet.dict is self.client.dict)
         self.assertEqual(packet.id, 15)
-        self.assertEqual(packet.secret, six.b('zeer geheim'))
+        self.assertEqual(packet.secret, b'zeer geheim')
 
     def testCreateAcctPacket(self):
         packet = self.client.CreateAcctPacket(id=15)
-        self.failUnless(isinstance(packet, AcctPacket))
-        self.failUnless(packet.dict is self.client.dict)
+        self.assertTrue(isinstance(packet, AcctPacket))
+        self.assertTrue(packet.dict is self.client.dict)
         self.assertEqual(packet.id, 15)
-        self.assertEqual(packet.secret, six.b('zeer geheim'))
+        self.assertEqual(packet.secret, b'zeer geheim')